+++ docs/changelog.rst
@@ -6,6 +6,7 @@ Changelog
 ------------------
 
 - #61 Fix Traceback in listing when fullname is different from sample's
+- #60 Fix patient's middlename is not displayed in samples listing
 - #59 Migrate Patient MRN/ID Sample Widgets to QuerySelect
 - #58 Add sample field validation for patient ID
 - #57 Patient samples listing and registration form